backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task Queue
redis==5.0.1
celery==5.3.4

# AI & ML
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.26
pinecone-client==2.2.4
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
PyPDF2==3.0.1
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2
